RDPP-7025: Export confluence into markdown (#21)
* Remove content and images for a clean slate.

* Tweak the linting rules.
Allow underscores in the names and item lists should have a single space.

* Disable the lists free form linting.
This plugin's linting was anoying. It didn't want a new lines after simple items in a list, but required one on multiline lists. After a discussion we decided to disable it.

* Disable the custom installation of dev packages from the build script.

* Bump versions.
This is not required, but might as well.. with the Linters pages gone now, lint the RN instead.

* Update content and images based on the exporter.